Allows device types to register themselves out-of-core
