+++ test/common_device_type.py
@@ -1,10 +1,13 @@
 import inspect
-import threading
 from functools import wraps
 import unittest
 import os
 import torch
-from common_utils import TestCase, TEST_WITH_ROCM, TEST_MKL, \
+import copy
+import re
+
+from common_device_type_base import DeviceTypeTestBase
+from common_utils import TEST_WITH_ROCM, TEST_MKL, \
     skipCUDANonDefaultStreamIf
 
 # Note: Generic Device-Type Testing
@@ -101,12 +104,11 @@ from common_utils import TestCase, TEST_WITH_ROCM, TEST_MKL, \
 #
 # To add a device type:
 #
-#   (1) Create a new "TestBase" extending DeviceTypeTestBase.
-#       See CPUTestBase and CUDATestBase below.
-#   (2) Define the "device_type" attribute of the base to be the
-#       appropriate string.
-#   (3) Add logic to this file that appends your base class to
-#       device_type_test_bases when your device type is available.
+#   (1) Create a file in this directory called "device_type_<name>.py"
+#   (2) Create a new "TestBase" extending DeviceTypeTestBase in that file.
+#       See CPUTestBase and CUDATestBase below for examples.
+#   (3) Implement a function get_test_base() in your file that returns
+#       this test base.
 #   (4) (Optional) Write setUpClass/tearDownClass class methods that
 #       instantiate dependencies (see MAGMA in CUDATestBase).
 #   (5) (Optional) Override the "instantiate_test" method for total
@@ -121,91 +123,11 @@ from common_utils import TestCase, TEST_WITH_ROCM, TEST_MKL, \
 # you should check if it's available and (if it is) add it to this list.
 device_type_test_bases = []
 
-
-class DeviceTypeTestBase(TestCase):
-    device_type = 'generic_device_type'
-
-    # Precision is a thread-local setting since it may be overriden per test
-    _tls = threading.local()
-    _tls.precision = TestCase.precision
-
-    @property
-    def precision(self):
-        return self._tls.precision
-
-    @precision.setter
-    def precision(self, prec):
-        self._tls.precision = prec
-
-    # Returns a string representing the device that single device tests should use.
-    # Note: single device tests use this device exclusively.
-    @classmethod
-    def get_primary_device(cls):
-        return cls.device_type
-
-    # Returns a list of strings representing all available devices of this
-    # device type. The primary device must be the first string in the list
-    # and the list must contain no duplicates.
-    # Note: UNSTABLE API. Will be replaced once PyTorch has a device generic
-    #   mechanism of acquiring all available devices.
-    @classmethod
-    def get_all_devices(cls):
-        return [cls.get_primary_device()]
-
-    # Returns the dtypes the test has requested.
-    # Prefers device-specific dtype specifications over generic ones.
-    @classmethod
-    def _get_dtypes(cls, test):
-        if not hasattr(test, 'dtypes'):
-            return None
-        return test.dtypes.get(cls.device_type, test.dtypes.get('all', None))
-
-    def _get_precision_override(self, test, dtype):
-        if not hasattr(test, 'precision_overrides'):
-            return self.precision
-        return test.precision_overrides.get(dtype, self.precision)
-
-    # Creates device-specific tests.
-    @classmethod
-    def instantiate_test(cls, name, test):
-        test_name = name + "_" + cls.device_type
-
-        dtypes = cls._get_dtypes(test)
-        if dtypes is None:  # Test has no dtype variants
-            assert not hasattr(cls, test_name), "Redefinition of test {0}".format(test_name)
-
-            @wraps(test)
-            def instantiated_test(self, test=test):
-                device_arg = cls.get_primary_device() if not hasattr(test, 'num_required_devices') else cls.get_all_devices()
-                return test(self, device_arg)
-
-            setattr(cls, test_name, instantiated_test)
-        else:  # Test has dtype variants
-            for dtype in dtypes:
-                dtype_str = str(dtype).split('.')[1]
-                dtype_test_name = test_name + "_" + dtype_str
-                assert not hasattr(cls, dtype_test_name), "Redefinition of test {0}".format(dtype_test_name)
-
-                @wraps(test)
-                def instantiated_test(self, test=test, dtype=dtype):
-                    device_arg = cls.get_primary_device() if not hasattr(test, 'num_required_devices') else cls.get_all_devices()
-                    # Sets precision and runs test
-                    # Note: precision is reset after the test is run
-                    guard_precision = self.precision
-                    try :
-                        self.precision = self._get_precision_override(test, dtype)
-                        result = test(self, device_arg, dtype)
-                    finally:
-                        self.precision = guard_precision
-
-                    return result
-
-                setattr(cls, dtype_test_name, instantiated_test)
-
-
 class CPUTestBase(DeviceTypeTestBase):
     device_type = 'cpu'
 
+
+PYTORCH_CUDA_MEMCHECK = os.getenv('PYTORCH_CUDA_MEMCHECK', '0') == '1'
 
 class CUDATestBase(DeviceTypeTestBase):
     device_type = 'cuda'
@@ -248,7 +170,18 @@ device_type_test_bases.append(CPUTestBase)
 if torch.cuda.is_available():
     device_type_test_bases.append(CUDATestBase)
 
-PYTORCH_CUDA_MEMCHECK = os.getenv('PYTORCH_CUDA_MEMCHECK', '0') == '1'
+# Finds other device types
+torch_test_dir = os.path.dirname(os.path.abspath(__file__))
+torch_test_files = [file for file in os.listdir(torch_test_dir) if os.path.isfile(os.path.join(torch_test_dir, file))]
+regex = re.compile('^device_type_(.+).py$')
+device_type_files = []
+for test_file in torch_test_files:
+    result = regex.match(test_file)
+    if result:
+        device_type_name = result.group(1)
+        module_name = result.group(0).split('.')[0]
+        device_type_module = __import__(module_name, globals(), locals(), [], 0)
+        device_type_test_bases.append(device_type_module.get_test_base())
 
 
 # Adds 'instantiated' device-specific test cases to the given scope.
@@ -292,7 +225,7 @@ def instantiate_device_type_tests(generic_test_class, scope, except_for=None):
                 assert inspect.isfunction(test), "Couldn't extract function from '{0}'".format(name)
 
                 # Instantiates the device-specific tests
-                device_type_test_class.instantiate_test(name, test)
+                device_type_test_class.instantiate_test(name, copy.deepcopy(test))
             else:  # Ports non-test member
                 assert not hasattr(device_type_test_class, name), "Redefinition of non-test member {0}".format(name)
 
@@ -440,9 +373,9 @@ def onlyOnCPUAndCUDA(fn):
 # functions like self.assertEqual().
 #
 # Note that self.precision is a scalar value, so if you require multiple
-# precisions (or are working with multiple dtypes) they should be specified
-# explicitly and computed using self.precision (e.g.
-# self.precision *2, max(1, self.precision)).
+# precisions (or are working with multiple dtypes) in your test then they should
+# be specified explicitly and computed using self.precision (e.g.
+# self.precision * 2, max(1, self.precision)).
 class precisionOverride(object):
 
     def __init__(self, d):
